Moves Qt related functions into tk_maya.__init__
This is to allow engines to shutdown and restart
without losing the dictionary of active widgets.
